Interpret keyframe interval value as milliseconds.
Change storage type from double to int.

+++ http_client/video_encoder.h
@@ -148,8 +148,8 @@ class VideoFrameCallbackInterface {
 };
 
 struct VpxConfig {
-  // Time between keyframes, in seconds.
-  double keyframe_interval;
+  // Time between keyframes, in milliseconds.
+  int keyframe_interval;
   // Video bitrate, in kilobits.
   int bitrate;
   // Video frame rate decimation factor.
